Resultados a decimal en migración para mostrar exactos

+++ database/migrations/2025_01_15_221426_variable_val_eval.php
@@ -30,7 +30,7 @@ return new class extends Migration
         Schema::create("evaluacion_result", function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger("evaluacionId");
-            $table->unsignedBigInteger("resultado");
+            $table->decimal("resultado");
             $table->string("status")->default("capturado"); // capturado, aprobado, rechazado
             $table->unsignedBigInteger("aprobadoPorId")->nullable();
             $table->date("fecha");
